Fix failing tests
The initial error was during compile time because of passing None for
the request `data` parameter. The docs were updated to reflect that
`None` must be parameterized like `None::<u8>`.

Having resolved that, there was a runtime error for the tests where one
of them would fail because the specified port was already used.
Presumably the tests run in parallel. The StackListener type was updated
to take a `port: u16` argument which it will attempt to listen on. This
isn't an ideal solution, and another issue will be filed to address it.

Resolves #1.